requirements.txt
flask==3.0.0
mistralai>=1.0.0
openai>=1.0.0
supabase>=2.0.0
gunicorn==21.2.0
Pillow>=10.2.0
python-dotenv==1.0.0
requests>=2.31.0
werkzeug>=3.0.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pdf2image==1.17.0
langdetect>=1.0.9